<FEAT>[<DOCS>]<Implementando novas classes para as funcoes de calculo>

+++ cap4/IExpressao.cs
@@ -2,6 +2,6 @@ namespace DesignPatterns2.cap4
 {
     public interface IExpressao
     {
-        int Avalia();
+        double Avalia();
     }
 }
